refactor(user): use new db structure

+++ src/main/java/me/jinwenjie/service/UserService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface UserService {
     List<User> list();
 
-    User get(Integer id);
+    User get(Long id);
 
     boolean create(User user);
 
     boolean update(User user);
 
-    boolean delete(Integer id);
+    boolean delete(Long id);
 }
